requirements.txt
gradio
numpy
Pillow
opencv-python
